requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.7.2
djangorestframework==3.13.1
goodlogging==1.0.1
gunicorn==20.1.0
Pillow==9.0.1
pytz==2018.4
rarfile==2.7
requests==2.6.0
sqlparse==0.4.2
whitenoise==6.0.0
